changed userid from int to long and removed lateinit usage from views

+++ features/promo/tokomember_seller_dashboard/src/main/java/com/tokopedia/tokomember_seller_dashboard/model/TmMemberListResponse.kt
@@ -48,7 +48,7 @@ data class UserCardMember (
     val id:String? = "",
     @Expose
     @SerializedName("userID")
-    val userID:Int?= 0,
+    val userID:Long?= 0,
     @Expose
     @SerializedName("referenceID")
     val referenceID:String?= "",
